refactor(rn): extract tab bar as a component
