+++ apps/mobile/src/components/ui/tabview/types.ts
@@ -1,0 +1,5 @@
+export type Tab = {
+  name: string
+  activeColor?: string
+  value: string
+}
