+++ apps/mobile/src/modules/discover/RecommendationListItem.tsx
