+++ apps/mobile/src/modules/discover/constants.ts
@@ -1,0 +1,6 @@
+export enum SearchType {
+  AGGREGATE = "aggregate",
+  RSS = "rss",
+  RSSHUB = "rsshub",
+  USER = "user",
+}
