+++ apps/mobile/src/modules/discover/ctx.tsx
@@ -2,9 +2,13 @@ import type { PrimitiveAtom } from "jotai"
 import { atom } from "jotai"
 import { createContext, useContext, useState } from "react"
 
+import { SearchType } from "./constants"
+
 interface DiscoverPageContextType {
   searchFocusedAtom: PrimitiveAtom<boolean>
   searchValueAtom: PrimitiveAtom<string>
+
+  searchTypeAtom: PrimitiveAtom<SearchType>
 }
 export const DiscoverPageContext = createContext<DiscoverPageContextType>(null!)
 
@@ -12,10 +16,11 @@ export const DiscoverPageProvider = ({ children }: { children: React.ReactNode }
   const [atomRefs] = useState((): DiscoverPageContextType => {
     const searchFocusedAtom = atom(true)
     const searchValueAtom = atom("")
-
+    const searchTypeAtom = atom(SearchType.AGGREGATE)
     return {
       searchFocusedAtom,
       searchValueAtom,
+      searchTypeAtom,
     }
   })
   return <DiscoverPageContext.Provider value={atomRefs}>{children}</DiscoverPageContext.Provider>
